pyproject.toml
[project]
name = "beermonopoly-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "blessed>=1.25.0",
    "cloudscraper25>=2.7.0",
    "cryptography>=43.0.3",
    "django-admin-shell>=2.0.1",
    "django-cors-headers>=4.5.0",
    "django-dirtyfields>=1.9.8",
    "django-extensions>=3.3.0",
    "django-filter>=25.2",
    "django-hosts>=6.0",
    "django-q2>=1.7.3",
    "django>=5.2,<5.3",
    "djangorestframework>=3.15.2",
    "drf-dynamic-fields>=0.4.0",
    "factory-boy>=3.3.3",
    "feedparser>=6.0.0",
    "filters>=1.3.2",
    "firebase-admin>=6.5.0",
    "freezegun>=1.5.1",
    "fuzzywuzzy>=0.18.0",
    "googlesearch-python>=1.2.5",
    "gunicorn>=23.0.0",
    "jupyter>=1.1.1",
    "lxml>=5.3.0",
    "pandas>=2.2.3",
    "psycopg-binary>=3.3.0",
    "psycopg>=3.3.0",
    "pyjwt>=2.9.0",
    "pytest-django>=4.9.0",
    "pytest>=9.0.1",
    "responses>=0.25.3",
    "ruff>=0.14.7",
    "sentry-sdk>=2.46.0",
    "urllib3>=2.2.3",
    "xmltodict>=0.14.2",
]

[tool.pyrefly]
search_path = ["api"]

[dependency-groups]
dev = [
    "codecov>=2.1.13",
    "coverage>=7.12.0",
    "django-filter-stubs>=0.1.3",
    "django-stubs>=5.2.0",
    "djangorestframework-stubs>=3.15.2",
    "types-requests>=2.32.0",
]
